extract staff widget
